- Migrated test case planner from int indices for test cases to BigInteger indices. - This should support arbitrary large search spaces. - Added interface to planner class hierarchy to aggregate literals and function application expressions. - Added NonExecutingPlanner to encapsulate planning activaties from NonExecutingCrasher.

+++ jcrasher-core/edu/gatech/cc/jcrasher/planner/ClassUnderTest.java
@@ -18,7 +18,7 @@ public interface ClassUnderTest<T> {
 	 * 
 	 * Precond: 0 <= planIndex < getPlanSpaceSize() Postcond: no side-effects
 	 */
-	public Block<?> getBlock(int planIndex);
+	public Block<?> getBlock(BigInteger planIndex);
   
   /**
    * @return the number of test methods available.
